MySQL SSL
Got a script written that generates the needed keys and certificates. The script is in SSL/generate_certs/gen_certs.bat and takes 3 arguments: the destination folder, the certificate algorithm, and the algorithm version. If none are specified it defaults to the current folder and RSA-2048. If only a folder is specified it still defaults to RSA-2048. The algorithm needs to be one of rsa, dsa, or ecdsa. For the algorithm version, rsa takes any integer >= 512 (but should be >= 1024), dsa takes any integer > 0 (but should be >= 1024), and ecdsa takes any one of a bunch of named elliptic curves such as prime256v1 (an alias for the default) or secp521r1. The default rsa and dsa versions are both 2048, and the default ecdsa version is secp256r1.
The script requires OpenSSL to be installed and its bin folder included in the PATH variable. It can be installed from https://slproweb.com/products/Win32OpenSSL.html
I'd recommend v1.1.1d.