JDBC MySQL SSL
Wrote a script (in SSL/generate_keystores/gen_keystores.bat) to generate the "truststore" and "keystore" files needed for the database connector to use SSL. The script takes between 1 and 4 arguments. The first argument is the password of the truststore. This is either the password of an existing truststore or the password for a new truststore. The second argument is the password of the keystore. Again this is either the password for the existing keystore or the new keystore. If only the 1st argument is given, the password for the keystore is the same as the password for the truststore.
The third argument is the folder that the certificates and keys generated by gen_certs.bat are stored at. If this is not provided, it defaults to the current directory. The fourth argument is the destination folder for the truststore and keystore. If this is not provided, it defaults to the current directory. To set the destination folder but not the certificates folder (i.e. the certificates are in the current directory and you don't want to type that out) just use . (a single period) as the certificates folder.