JDBC SSL Configuration
Added 2 new properties to SmallTownShipsConfig.xml, the truststore password and the keystore password, and configured MySQLHandler.java to use an SSL connection. This requires that the "truststore" and "keystore" generated by gen_keystores.bat are both in a folder named SSL that is in the same directory as SmallTownShipsConfig.xml. I've also included an example XML file that uses the passwords set for the truststore and keystore that have already been generated using the certificates and keys in the SSL folder.