Add Session Management
There are a lot of changes here.

**Changes:**

The `LoginHandler` class was removed and replaced with a `User` class, which allows for more Object-Oriented programming throughout the project. Fixes #11

The `User` class has getters and setters for many properties of the user (name, email, etc.,). `loggedInUser` fetches the currently authenticated user from the session, if there is one.

Any instance of `LoginHandler` was updated to use the new User model.

Some of the SQL procedures were changed, including adding one, `User_With_Token` which returns a user by their login token. The login token is a randomly generated integer that uniqiuely identifies a user's session.

The `Try_Login` procedure now sets the login token, if the user is found.

The `Update_Login` procedure was removed.

The `Logout_User` procedure sets the users login token to null.

I added a new `callProcedure` method in MySQLHandler that doesn't require the number of arguments to be passed in, nor require the array of parameters to be strings. It also [allows an arbitrary number of arguments](https://docs.oracle.com/javase/tutorial/java/javaOO/arguments.html#varargs). In the future, I'd like to replace all instances of callProcedure with the new one.

**Please make sure this runs on your machine before I merge it into master.**

To do this, switch to the `add-session-management` branch on GitHub Desktop. [Instructions here](https://help.github.com/en/desktop/contributing-to-projects/switching-between-branches).

Thanks for coming to my TED talk.