Added password hashing and credit card number encryption
Passwords are now stored as a salt and hash, and credit card information is encrypted using the user's password to generate a key. It isn't anything fancy but it took awhile to get everything worked in, especially pre-defining the manager hashed password. The password is still "ims" but its stored as a hash.