Improved password hashing function
Made the SHA-256 password hashing more compliant with the FIPS standard. Still uses 5000 rounds instead of 64 rounds, but the round constants for these rounds are now generated from the first 32 bits of the fractional part of the cube root of the first 5000 prime numbers. This is consistent with the FIPS standard of SHA-256, which uses the first 32 bits of the fractional part of the cube root of the first 64 prime numbers as the round constants.

This is technically NOT a FIPS-compliant implementation of SHA-256, however the only difference is the fact that this implementation uses 5000 rounds of compression, rather than simply 64, and the round constants are expanded to the first 32 bits of the fractional part of the cube root of the first 5000 prime numbers, rather than simply the first 64 prime numbers.

TL;DR The password hash function now uses a unique round constant for each round, rather than recycling the 64 rounds constants from the FIPS standard.

Also updated the password for the manager account in the database initialization script to comply with the updated hashing algorithm.

** Previous reverted commit used the wrong source of random number to generate the salt **
(probably should have just done a separate commit to fix it...)